fix: some situations that prompt wrongly or not redraw
When using `v` key to enter `visual` mode and then use `iw` to select
in a word, the keymap will be changed to  vicmd, this would cause a
wrong prompt, since it still in visual mode. Also when you select a
world and then use key `c/s/d`, the prompt doesn't redraw, and so on.