fix: the weird confict with the `starship` plugin
When there was only the `starship` and no the `zsh-autosuggest`
plugin was loaded, there will cause a weird behaviour when you
enter and exit insert mode, it's that the prompt line will move
up and overwrite the previous lines. This was caused by multiple
calling the `reset-prompt` widget when selecting the keymap, the
`startship` will call the `reset-prompt` widget for refreshing
the prompt.