perf: auto restore cursor to user default (#35)
When we start a program (e.g. vim, bash, etc.) from the
command line, the cursor style is inherited by other
programs, so that we need to reset the cursor style to
default before executing a command and set the custom
style again when the command exits. This way makes any
other interactive CLI application would not be affected
by it.