feat: new option ZVM_TERM to set term type (#85)
In order to handle the escape sequences properly, we should give a
proper value to the TERM. As it's important for terminal emulator
to handle escape sequences properly, so we can not casually change
it.

Here we provide an option ZVM_TERM for users to configurate and no
need to change the global TERM , its default value is the same as
TERM. For some reasons, you can change it to let this plugin work
properly.