Add note about timeouts in surround guide
I was pretty confused why `va"` was working fine, but `va(` was not. Turns out I was just too slow typing that parens. Adding a note to spare other slow typists the same confusion.
Also, it would be great if the `b` alias for `(` worked too. I believe this is enabled by default (at least it is in neovim). This lets commands such `vab` behave as `va(`.